# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AmbiguityChoice": {
        "properties": {
          "item_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Id"
          },
          "record_id": {
            "title": "Record Id",
            "type": "integer"
          }
        },
        "required": [
          "record_id"
        ],
        "title": "AmbiguityChoice",
        "type": "object"
      },
...
      "Body_preview_api_import_goodreads_preview_post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_preview_api_import_goodreads_preview_post",
        "type": "object"
      },
...
      "CalibrePreviewBody": {
        "properties": {
          "library_path": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Library Path",
            "type": "string"
          }
        },
        "required": [
          "library_path"
        ],
        "title": "CalibrePreviewBody",
        "type": "object"
      },
      "CommitBody": {
        "properties": {
          "batch_id": {
            "title": "Batch Id",
            "type": "string"
          },
          "choices": {
            "items": {
              "$ref": "#/components/schemas/AmbiguityChoice"
            },
            "title": "Choices",
            "type": "array"
          }
        },
        "required": [
          "batch_id"
        ],
        "title": "CommitBody",
        "type": "object"
      },
      "CommitResponse": {
        "properties": {
          "batch_id": {
            "title": "Batch Id",
            "type": "string"
          },
          "created_entries": {
            "title": "Created Entries",
            "type": "integer"
          },
          "created_items": {
            "title": "Created Items",
            "type": "integer"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "unchanged_entries": {
            "title": "Unchanged Entries",
            "type": "integer"
          },
          "unsorted_entries": {
            "title": "Unsorted Entries",
            "type": "integer"
          }
        },
        "required": [
          "batch_id",
          "state",
          "created_items",
          "created_entries",
          "unchanged_entries",
          "unsorted_entries"
        ],
        "title": "CommitResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobProgressResponse": {
        "properties": {
          "attempts": {
            "default": 0,
            "title": "Attempts",
            "type": "integer"
          },
          "batch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "progress": {
            "additionalProperties": true,
            "title": "Progress",
            "type": "object"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "state",
          "created_at"
        ],
        "title": "JobProgressResponse",
        "type": "object"
      },
...
      "PreviewResponse": {
        "properties": {
          "batch_id": {
            "title": "Batch Id",
            "type": "string"
          },
          "fingerprint": {
            "title": "Fingerprint",
            "type": "string"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/ImportRecordResponse"
            },
            "title": "Records",
            "type": "array"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "summary": {
            "$ref": "#/components/schemas/PreviewSummary"
          }
        },
        "required": [
          "batch_id",
          "fingerprint",
          "state",
          "summary",
          "records"
        ],
        "title": "PreviewResponse",
        "type": "object"
      },
      "PreviewSummary": {
        "properties": {
          "ambiguous": {
            "title": "Ambiguous",
            "type": "integer"
          },
          "errors": {
            "title": "Errors",
            "type": "integer"
          },
          "ready": {
            "title": "Ready",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "ready",
          "errors",
          "ambiguous"
        ],
        "title": "PreviewSummary",
        "type": "object"
      },
...
      "SearchCandidateResponse": {
        "properties": {
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "creators": {
            "items": {
              "type": "string"
            },
            "title": "Creators",
            "type": "array"
          },
          "credit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credit"
          },
          "identifiers": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Identifiers",
            "type": "object"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "original_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Year"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_refs": {
            "items": {
              "$ref": "#/components/schemas/SourceRefResponse"
            },
            "title": "Source Refs",
            "type": "array"
          },
          "subtitle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "source",
          "source_id",
          "source_refs",
          "title",
          "subtitle",
          "creators",
          "credit",
          "year",
          "original_year",
          "cover_url",
          "identifiers",
          "language",
          "metadata"
        ],
        "title": "SearchCandidateResponse",
        "type": "object"
      },
...
      "SourceRefResponse": {
        "properties": {
          "source": {
            "title": "Source",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          }
        },
        "required": [
          "source",
          "source_id"
        ],
        "title": "SourceRefResponse",
        "type": "object"
      },
...
      "UndoEffectSummary": {
        "properties": {
          "batch_id": {
            "title": "Batch Id",
            "type": "string"
          },
          "retained": {
            "title": "Retained",
            "type": "integer"
          },
          "retained_items": {
            "default": 0,
            "title": "Retained Items",
            "type": "integer"
          },
          "reverted": {
            "title": "Reverted",
            "type": "integer"
          },
          "reverted_entries": {
            "default": 0,
            "title": "Reverted Entries",
            "type": "integer"
          },
          "reverted_items": {
            "default": 0,
            "title": "Reverted Items",
            "type": "integer"
          },
          "skipped": {
            "title": "Skipped",
            "type": "integer"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "batch_id",
          "state",
          "reverted",
          "retained",
          "skipped"
        ],
        "title": "UndoEffectSummary",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Akasha Book Tracker",
    "version": "1.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/import/batches/{batch_id}": {
      "delete": {
        "operationId": "undo_batch_api_import_batches__batch_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UndoEffectSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Undo Batch",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/import/calibre/commit": {
      "post": {
        "operationId": "calibre_commit_api_import_calibre_commit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommitBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calibre Commit",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/import/calibre/preview": {
      "post": {
        "operationId": "calibre_preview_api_import_calibre_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalibrePreviewBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calibre Preview",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/import/goodreads/commit": {
      "post": {
        "operationId": "commit_api_import_goodreads_commit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommitBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Commit",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/import/goodreads/preview": {
      "post": {
        "operationId": "preview_api_import_goodreads_preview_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_preview_api_import_goodreads_preview_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/import/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_progress_api_import_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobProgressResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job Progress",
        "tags": [
          "imports"
        ]
      }
    },
...
    "/api/search": {
      "get": {
        "operationId": "search_api_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 300,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SearchCandidateResponse"
                  },
                  "title": "Response Search Api Search Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search"
      }
    },
    "/api/search/resolve": {
      "get": {
        "operationId": "resolve_api_search_resolve_get",
        "parameters": [
          {
            "in": "query",
            "name": "url",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Url",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SearchCandidateResponse"
                  },
                  "title": "Response Resolve Api Search Resolve Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve"
      }
    },
...
  }
}